refactor/#64: Auth에 UseCase 패턴 적용
