+++ src/auth/application/dto/response/jwt-token.response.dto.ts
@@ -1,0 +1,13 @@
+export class JwtTokenResponseDto {
+  accessToken: string;
+  refreshToken: string;
+
+  constructor(accessToken: string, refreshToken: string) {
+    this.accessToken = accessToken;
+    this.refreshToken = refreshToken;
+  }
+
+  static of(accessToken: string, refreshToken: string): JwtTokenResponseDto {
+    return new JwtTokenResponseDto(accessToken, refreshToken);
+  }
+}
