+++ src/auth/application/usecase/change-password.usecase.ts
@@ -1,0 +1,5 @@
+import { UpdatePasswordRequestDto } from '../dto/request/update-password.request.dto';
+
+export interface ChangePasswordUseCase {
+  execute(userId: number, changePasswordDto: UpdatePasswordRequestDto,): Promise<void>;
+}
