+++ src/auth/application/usecase/logout.usecase.ts
@@ -1,0 +1,3 @@
+export interface LogoutUseCase {
+  execute(refreshToken: string): Promise<void>;
+}
