+++ src/auth/application/usecase/read-auth-brief.usecase.ts
@@ -1,0 +1,5 @@
+import { ReadAuthBriefResponseDto } from '../dto/response/read-auth-brief.response.dto';
+
+export interface ReadAuthBriefUsecase {
+  execute(accessToken: string): Promise<ReadAuthBriefResponseDto>;
+}
