+++ src/auth/application/usecase/reissue-jwt.usecase.ts
@@ -1,0 +1,5 @@
+import { JwtTokenResponseDto } from '../dto/response/jwt-token.response.dto';
+
+export interface ReissueJwtUseCase {
+  execute(refreshToken?: string): Promise<JwtTokenResponseDto>;
+}
