+++ src/auth/application/usecase/sign-up-admin.usecase.ts
@@ -1,0 +1,5 @@
+import { SignUpAdminRequestDto } from '../dto/request/sign-up-admin.request.dto';
+
+export interface SignUpAdminUseCase {
+  execute(requestDto: SignUpAdminRequestDto): Promise<void>;
+}
