+++ src/auth/application/usecase/sign-up.usecase.ts
@@ -1,0 +1,3 @@
+export interface SignUpUseCase {
+  execute(userId: number, file: Express.Multer.File): Promise<void>;
+}
