--- src/auth/auth.module.ts
@@ -1,26 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-import { JwtStrategy } from './jwt.strategy';
-import { JwtAuthGuard } from '../core/guards/jwt-auth.guard';
-import { JwtModule } from '@nestjs/jwt';
-import { PassportModule } from '@nestjs/passport';
-import { UserModule } from '../user/application/modules/user.module';
-import { DataSource } from 'typeorm';
-import { UserRepositoryImpl } from '../user/repository/user.repository.impl';
-
-@Module({
-  imports: [
-    UserModule,
-    PassportModule,
-    JwtModule.register({
-      secret: process.env.JWT_SECRET,
-      signOptions: { expiresIn: '1h' },
-    }),
-  ],
-  providers: [AuthService, JwtStrategy, JwtAuthGuard, UserRepositoryImpl],
-  controllers: [AuthController],
-})
-export class AuthModule {
-  constructor(private dataSource: DataSource) {}
-}
