--- src/auth/dto/response/reissue-jwt-token.response.dto.ts
@@ -1,13 +1,0 @@
-export class ReissueJwtTokenResponseDto {
-  accessToken: string;
-  refreshToken: string;
-
-  constructor(accessToken: string, refreshToken: string) {
-    this.accessToken = accessToken;
-    this.refreshToken = refreshToken;
-  }
-
-  static of(accessToken: string, refreshToken: string): ReissueJwtTokenResponseDto {
-    return new ReissueJwtTokenResponseDto(accessToken, refreshToken);
-  }
-}
