+++ src/core/strategies/jwt.strategy.ts
